+++ dataline-scheduler/src/main/java/io/dataline/scheduler/Scheduler.java
@@ -1,302 +1,32 @@
 package io.dataline.scheduler;
 
-import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ThreadFactoryBuilder;
-import io.dataline.api.model.ConnectionRead;
-import io.dataline.api.model.ConnectionSchedule;
-import io.dataline.api.model.ConnectionStatus;
-import io.dataline.api.model.Job;
-import io.dataline.db.DatabaseHelper;
-import io.dataline.workers.JobStatus;
-import io.dataline.workers.OutputAndStatus;
-import io.dataline.workers.Worker;
-import java.sql.SQLException;
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
 import java.util.concurrent.*;
 
 import org.apache.commons.dbcp2.BasicDataSource;
-import org.jooq.Record;
-import org.jooq.Result;
-import org.jooq.Table;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import static org.jooq.impl.DSL.field;
-import static org.jooq.impl.DSL.table;
-
-// todo: handle the fact these arent connection ids anymore
-// todo: split into separate files
-// todo: add comment blurb to describe purpose of this class
+/**
+ * The Scheduler is responsible for finding new scheduled jobs that need to be run and to launch
+ * them. The current implementation uses a thread pool on the scheduler's machine to launch the
+ * jobs. One thread is reserved for the job submitter, which is responsible for finding and
+ * launching new jobs.
+ */
 public class Scheduler {
-  private static final Logger LOGGER = LoggerFactory.getLogger(Scheduler.class);
-
   private static final int MAX_WORKERS = 4;
   private static final int JOB_SUBMITTER_THREAD = 1;
   private static final int MAX_THREADS = MAX_WORKERS + JOB_SUBMITTER_THREAD;
   private static final ThreadFactory THREAD_FACTORY =
       new ThreadFactoryBuilder().setNameFormat("scheduler-%d").build();
-  private final BasicDataSource connectionPool;
 
-  // todo: make endpoints in the API for reading job state
+  private final BasicDataSource connectionPool;
 
   public Scheduler(BasicDataSource connectionPool) {
     this.connectionPool = connectionPool;
   }
 
-  private static class SchedulerShutdownHookThread extends Thread {
-    private ExecutorService threadPool;
-
-    public SchedulerShutdownHookThread(ExecutorService threadPool) {
-      this.threadPool = threadPool;
-    }
-
-    @Override
-    public void run() {
-      // todo: attempt graceful shutdown of workers
-
-      threadPool.shutdown();
-
-      try {
-        if (!threadPool.awaitTermination(30, TimeUnit.SECONDS)) {
-          LOGGER.error("Unable to kill worker threads by shutdown timeout.");
-        }
-      } catch (InterruptedException e) {
-        LOGGER.error("Wait for graceful worker thread shutdown interrupted.", e);
-      }
-    }
-  }
-
-  // todo: provide blocking worker execution
-
-  private static class EchoWorker implements Worker<String> {
-    public EchoWorker() {}
-
-    @Override
-    public OutputAndStatus<String> run() {
-      LOGGER.info("Hello World");
-      return new OutputAndStatus<>(JobStatus.SUCCESSFUL, "echoed");
-    }
-
-    @Override
-    public void cancel() {
-      // no-op
-    }
-  }
-
-  // todo: test constructing this
-  private static class WorkerWrapper<T> implements Runnable {
-    private final long jobId;
-    private final Worker<T> worker;
-    private BasicDataSource connectionPool;
-
-    public WorkerWrapper(long jobId, Worker<T> worker, BasicDataSource connectionPool) {
-      this.jobId = jobId;
-      this.worker = worker;
-      this.connectionPool = connectionPool;
-    }
-
-    @Override
-    public void run() {
-      LOGGER.info("executing worker wrapper...");
-      try {
-        setJobStatus(connectionPool, jobId, Job.StatusEnum.RUNNING);
-        // todo: use attempt here
-
-        OutputAndStatus<T> outputAndStatus = worker.run();
-
-        switch (outputAndStatus.status) {
-          case FAILED:
-            setJobStatus(connectionPool, jobId, Job.StatusEnum.FAILED);
-            break;
-          case SUCCESSFUL:
-            setJobStatus(connectionPool, jobId, Job.StatusEnum.COMPLETED);
-            break;
-        }
-
-        // todo: propagate output
-        LOGGER.info("output " + outputAndStatus.output.toString());
-      } catch (Exception e) {
-        LOGGER.error("worker error", e);
-        setJobStatus(connectionPool, jobId, Job.StatusEnum.FAILED);
-      }
-    }
-  }
-
-  private static UUID CONNECTION_ID = UUID.randomUUID();
-
-  private static Set<ConnectionRead> getAllActiveConnections() {
-    // todo: get this from the api or a helper
-    ConnectionSchedule testConnectionSchedule = new ConnectionSchedule();
-    testConnectionSchedule.setUnits(1);
-    testConnectionSchedule.setTimeUnit(ConnectionSchedule.TimeUnitEnum.MINUTES);
-
-    ConnectionRead testConnection = new ConnectionRead();
-    testConnection.setName("echo-connection");
-    testConnection.setConnectionId(CONNECTION_ID);
-    testConnection.setSourceImplementationId(UUID.randomUUID());
-    testConnection.setDestinationImplementationId(UUID.randomUUID());
-    testConnection.setSchedule(testConnectionSchedule);
-    testConnection.setStatus(ConnectionStatus.ACTIVE);
-
-    return Sets.newHashSet(testConnection);
-  }
-
-  private static void setJobStatus(
-      BasicDataSource connectionPool, long jobId, Job.StatusEnum status) {
-    LOGGER.info("setting job status to " + status + " for job " + jobId);
-    try {
-      DatabaseHelper.query(
-          connectionPool,
-          ctx ->
-              ctx.execute(
-                  "UPDATE jobs SET status = CAST(? as JOB_STATUS) WHERE id = ?",
-                  status.toString().toLowerCase(),
-                  jobId));
-    } catch (SQLException e) {
-      LOGGER.error("sql", e);
-      throw new RuntimeException(e); // todo: actually handle this error
-    }
-  }
-
-  // todo: have discovery and oneoffs be blocking (how to handle logs)
-
-  private static class JobSubmitterThread implements Runnable {
-    private final ExecutorService threadPool;
-    private BasicDataSource connectionPool;
-
-    public JobSubmitterThread(ExecutorService threadPool, BasicDataSource connectionPool) {
-      this.threadPool = threadPool;
-      this.connectionPool = connectionPool;
-    }
-
-    @Override
-    public void run() {
-      try {
-        while (true) {
-          LOGGER.info("running job-submitter...");
-          Set<ConnectionRead> activeConnections = getAllActiveConnections();
-          for (ConnectionRead connection : activeConnections) {
-            Optional<Job> lastJob =
-                DatabaseHelper.query(
-                    connectionPool,
-                    ctx -> {
-                      Optional<Record> jobEntryOptional =
-                          ctx
-                              .fetch(
-                                  "SELECT * FROM jobs WHERE scope = ? AND CAST(status AS VARCHAR) <> ? ORDER BY created_at DESC LIMIT 1",
-                                  connection.getConnectionId().toString(),
-                                  Job.StatusEnum.CANCELLED.toString().toLowerCase())
-                              .stream()
-                              .findFirst();
-
-                      if (jobEntryOptional.isPresent()) {
-                        Record jobEntry = jobEntryOptional.get();
-                        Job job = new Job();
-                        job.setId(jobEntry.getValue("id", Long.class));
-                        job.setConnection(connection);
-                        job.setCreatedAt(jobEntry.getValue("created_at", LocalDateTime.class).toEpochSecond(ZoneOffset.UTC));
-                        job.setStartedAt(jobEntry.getValue("started_at", LocalDateTime.class).toEpochSecond(ZoneOffset.UTC));
-                        job.setStatus(
-                            Job.StatusEnum.fromValue(jobEntry.getValue("status", String.class)));
-                        job.setUpdatedAt(jobEntry.getValue("updated_at", LocalDateTime.class).toEpochSecond(ZoneOffset.UTC));
-                        return Optional.of(job);
-                      } else {
-                        return Optional.empty();
-                      }
-                    });
-
-            if (lastJob.isEmpty()) {
-              LOGGER.info("creating pending job for empty last job...");
-              createPendingJob(connectionPool, connection.getConnectionId(), "{}");
-              LOGGER.info("created pending job for empty last job...");
-            } else {
-              Job job = lastJob.get();
-
-              switch (job.getStatus()) {
-                case COMPLETED:
-                  ConnectionSchedule schedule = connection.getSchedule();
-                  long nextRunStart = job.getUpdatedAt() + getIntervalInSeconds(schedule);
-                  if (nextRunStart < Instant.now().getEpochSecond()) {
-                    createPendingJob(connectionPool, job.getConnection().getConnectionId(), "{}");
-                  }
-                  break; // executes on next iteration
-                case PENDING:
-                case FAILED: // infinite retries for now
-                  // todo: select kind of worker object to create
-                  LOGGER.info("submitting job to thread pool...");
-                  threadPool.submit(
-                      new WorkerWrapper<>(job.getId(), new EchoWorker(), connectionPool));
-                case RUNNING:
-                case CANCELLED:
-                  //  no-op
-                  break;
-              }
-            }
-          }
-
-          Thread.sleep(1000);
-        }
-
-      } catch (Throwable e) {
-        LOGGER.error("sql", e);
-      }
-    }
-
-    private static void createPendingJob(BasicDataSource connectionPool, UUID connectionId, String configJson) {
-      LOGGER.info("creating pending job for connection: " + connectionId.toString());
-      LocalDateTime now = LocalDateTime.ofInstant(Instant.now(), ZoneOffset.UTC);
-      try {
-        DatabaseHelper.query(
-            connectionPool,
-            ctx ->
-                ctx.execute(
-                    "INSERT INTO jobs VALUES(DEFAULT, ?, ?, ?, ?, CAST(? AS JOB_STATUS), CAST(? as JSONB), ?, ?, ?)",
-                    connectionId.toString(),
-                    now,
-                    now,
-                    now,
-                    Job.StatusEnum.PENDING.toString(),
-                    configJson,
-                    null, // todo: output
-                    "", // todo: assign stdout
-                    "") // todo: assign stderr
-            );
-      } catch (SQLException e) {
-        LOGGER.error("sql", e);
-        e.printStackTrace();
-      }
-    }
-
-    // todo: align daily with calendar days, etc instead 24 hours after last run completion (drift
-    // will be confusing)
-    private static Long getIntervalInSeconds(ConnectionSchedule schedule) {
-      switch (schedule.getTimeUnit()) {
-        case MINUTES:
-          return TimeUnit.MINUTES.toSeconds(1) * schedule.getUnits().longValue();
-        case HOURS:
-          return TimeUnit.HOURS.toSeconds(1) * schedule.getUnits().longValue();
-        case DAYS:
-          return TimeUnit.DAYS.toSeconds(1) * schedule.getUnits().longValue();
-        case WEEKS:
-          return TimeUnit.DAYS.toSeconds(1) * 7 * schedule.getUnits().longValue();
-        case MONTHS:
-          return TimeUnit.DAYS.toSeconds(1) * 30 * schedule.getUnits().longValue();
-        default:
-          throw new RuntimeException("unhandled enum time unit: " + schedule.getTimeUnit());
-      }
-    }
-  }
-
   public void start() {
     ExecutorService threadPool = Executors.newFixedThreadPool(MAX_THREADS, THREAD_FACTORY);
-
     threadPool.submit(new JobSubmitterThread(threadPool, connectionPool));
-
-    Runtime.getRuntime().addShutdownHook(new SchedulerShutdownHookThread(threadPool));
+    Runtime.getRuntime().addShutdownHook(new SchedulerShutdownThread(threadPool));
   }
 }
